% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_china.R
\docType{data}
\name{china_2002_122}
\alias{china_2002_122}
\title{Input-output table for China, 2002, 122 sectors}
\format{
A data frame with 129 rows and 139 columns.
}
\source{
\href{https://data.stats.gov.cn/ifnormal.htm?u=/files/html/quickSearch/trcc/trcc01.html&h=740}{National Bureau of Statistics of China}
}
\usage{
data(china_2002_122)
}
\description{
The national input-output table of China for 2002, covering 122 sectors.
    Data are calculated at producers' prices in 2002.
    Values are denoted in 10 thousand Chinese Yuan (CNY).
}
\keyword{datasets}
