% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_china.R
\docType{data}
\name{china_2007_135}
\alias{china_2007_135}
\title{Input-output table for China, 2007, 135 sectors}
\format{
A data frame with 142 rows and 152 columns.
}
\source{
\href{https://www.stats.gov.cn}{National Bureau of Statistics of China}
}
\usage{
data(china_2007_135)
}
\description{
The national input-output table of China for 2007, covering 135 sectors.
    Data are calculated at producers' prices in 2007.
    Values are denoted in 10 thousand Chinese Yuan (CNY).
}
\keyword{datasets}
