% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_china.R
\docType{data}
\name{china_2010_41}
\alias{china_2010_41}
\title{Input-output table for China, 2010, 41 sectors}
\format{
A data frame with 48 rows and 57 columns.
}
\source{
\href{https://data.stats.gov.cn/ifnormal.htm?u=/files/html/quickSearch/trcc/trcc01.html&h=740}{National Bureau of Statistics of China}
}
\usage{
data(china_2010_41)
}
\description{
The national input-output table of China for 2010, covering 41 sectors.
    Data are calculated at producers' prices in 2010.
    Values are denoted in 10 thousand Chinese Yuan (CNY).
}
\keyword{datasets}
