% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_china.R
\docType{data}
\name{china_2012_139}
\alias{china_2012_139}
\title{Input-output table for China, 2012, 139 sectors}
\format{
A data frame with 146 rows and 155 columns.
}
\source{
\href{https://data.stats.gov.cn/ifnormal.htm?u=/files/html/quickSearch/trcc/trcc01.html&h=740}{National Bureau of Statistics of China}
}
\usage{
data(china_2012_139)
}
\description{
The national input-output table of China for 2012, covering 139 sectors.
    Data are calculated at producers' prices in 2012.
    Values are denoted in 10 thousand Chinese Yuan (CNY).
}
\keyword{datasets}
