% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_china.R
\docType{data}
\name{china_2015_42}
\alias{china_2015_42}
\title{Input-output table for China, 2015, 42 sectors}
\format{
A data frame with 49 rows and 59 columns.
}
\source{
\href{https://data.stats.gov.cn/ifnormal.htm?u=/files/html/quickSearch/trcc/trcc01.html&h=740}{National Bureau of Statistics of China}
}
\usage{
data(china_2015_42)
}
\description{
The national input-output table of China for 2015, covering 42 sectors.
    Data are calculated at producers' prices in 2015.
    Values are denoted in 10 thousand Chinese Yuan (CNY).
}
\keyword{datasets}
