% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_china.R
\docType{data}
\name{china_2017_149}
\alias{china_2017_149}
\title{Input-output table for China, 2017, 149 sectors}
\format{
A data frame with 156 rows and 165 columns.
}
\source{
\href{https://data.stats.gov.cn/ifnormal.htm?u=/files/html/quickSearch/trcc/trcc01.html&h=740}{National Bureau of Statistics of China}
}
\usage{
data(china_2017_149)
}
\description{
The national input-output table of China for 2017, covering 149 sectors.
    Data are calculated at producers' prices in 2017.
    Values are denoted in 10 thousand Chinese Yuan (CNY).
}
\keyword{datasets}
