% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_china.R
\docType{data}
\name{china_2018_153}
\alias{china_2018_153}
\title{Input-output table for China, 2018, 153 sectors}
\format{
A data frame with 160 rows and 169 columns.
}
\source{
\href{https://data.stats.gov.cn/ifnormal.htm?u=/files/html/quickSearch/trcc/trcc01.html&h=740}{National Bureau of Statistics of China}
}
\usage{
data(china_2018_153)
}
\description{
The national input-output table of China for 2018, covering 153 sectors.
    Data are calculated at producers' prices in 2018.
    Values are denoted in 10 thousand Chinese Yuan (CNY).
}
\keyword{datasets}
