% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_china.R
\docType{data}
\name{china_2018_42}
\alias{china_2018_42}
\title{Input-output table for China, 2018, 42 sectors}
\format{
A data frame with 91 rows and 53 columns.
}
\source{
\href{https://www.stats.gov.cn}{National Bureau of Statistics of China}
}
\usage{
data(china_2018_42)
}
\description{
The national input-output table of China for 2018, covering 42 sectors.
    Values are denoted in 10 thousand Chinese Yuan (CNY).
}
\keyword{datasets}
