% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_China.R
\docType{data}
\name{china_2020_42}
\alias{china_2020_42}
\title{Input-output table for China, 2020, 42 sectors}
\format{
A data frame with 91 rows and 53 columns.
}
\source{
\href{https://data.stats.gov.cn/ifnormal.htm?u=/files/html/quickSearch/trcc/trcc01.html&h=740}{National Bureau of Statistics of China}
}
\usage{
data(china_2020_42)
}
\description{
The national input-output table of China for 2020, covering 42 sectors.
    Values are denoted in 10 thousand Chinese Yuan (CNY).
}
\keyword{datasets}
