% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_china.R
\docType{data}
\name{china_employment}
\alias{china_employment}
\title{Sectoral employment data for China, 1990-2018}
\format{
A data frame with 115 rows and 32 columns.
}
\usage{
data(china_employment)
}
\description{
Number of employed persons (in the unit of 10 thousands) in China 1990-2018, by sector.
}
\references{
Wang, Y., Jia, X. and Wang, C. (2021).
Research on Sectoral Employment Accounting of China.
\emph{Statistical Research}, 38(12), 3--18.
\href{https://tjyj.stats.gov.cn/EN/Y2021/V38/I12/3}{URL}
}
\keyword{datasets}
