% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_WIOD.R
\name{data_WIOD}
\alias{data_WIOD}
\title{Get data from the World Input-Output Database (WIOD)}
\source{
\href{https://www.rug.nl/ggdc/valuechain/wiod/}{World Input-Output Database}
}
\usage{
data_WIOD(version = "2016 Release", year)
}
\arguments{
\item{version}{is the specific release of WIOD, either "2013 Release" or "2016 Release".}

\item{year}{is the specific year of the data.}
}
\description{
Download the data from \href{https://github.com/Carol-seven/ionet_ext}{ionet_ext},
an extension repository for storing data in the R package ionet,
and store it in the package directory.
If the data has already been downloaded, a message will be displayed.
}
\note{
The size of the data may exceed the capacity of the internet connection,
potentially limiting the download speed.
}
