% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_OECD21ed_1995to2018.R
\docType{data}
\name{OECD21ed_JPN_1995to2018}
\alias{OECD21ed_JPN_1995to2018}
\title{Input-output tables of the OECD 2021 edition for Japan, 1995--2018, 45 sectors}
\format{
A list containing 24 data frames, each with 50 rows and 56 columns.
}
\source{
\href{https://www.oecd.org/sti/ind/input-outputtables.htm}{Input-Output Tables (IOTs) 2021 ed. - OECD}
}
\usage{
data(OECD21ed_JPN_1995to2018)
}
\description{
The national input-output tables of Japan for the period from 1995 to 2018, covering 45 sectors.
    Data are calculated in current prices.
    Values are denoted in millions of US dollars (USD).
}
\keyword{datasets}
