% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betweenness.R
\name{btw}
\alias{btw}
\title{Betweenness centrality}
\usage{
btw(
  adjmat,
  gross,
  prior = rep(1, dim(adjmat)[1]),
  alpha = 1,
  type = c("consumption", "distribution")
)
}
\arguments{
\item{adjmat}{is a square matrix of intermediate transactions in the input-output table.}

\item{gross}{is the total input/output vector.}

\item{prior}{is the vector containing node-specific prior information.}

\item{alpha}{is the tuning parameter controlling the weights for SP strength and
prior information.}

\item{type}{which type of SP to calculate: "consumption" or "distribution".}
}
\value{
a list of the betweeness centrality (vector), associated SP (matrix recording
the intermediate vertices) and SP strength (matrix) for all vertices.
}
\description{
Compute the betweenness centrality with priors based on strongest path.
}
