% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{btw}
\alias{btw}
\title{Betweenness centrality}
\usage{
btw(adjmat, gross, prior, alpha = 1, type = "consumption")
}
\arguments{
\item{adjmat}{The matrix of intermediate transactions in the input-output table,
a n-by-n matrix.}

\item{gross}{The gross input/output, a n-by-1 vector.}

\item{prior}{The node-specific prior information, a n-by-1 vector.}

\item{alpha}{The tuning parameter controlling the weights for SP strength and
prior information. The default setting is 1.}

\item{type}{Which type of SP to calculate: "consumption"/"pull" (default) or "distribution"/"push".}
}
\value{
Lists of betweeness score, associated SPs, SP distance and SP strength.
}
\description{
Compute the betweenness centrality with prior information based on strongest path (SP).
}
\references{
Xiao, S., Yan, J. and Zhang, P. (2022).
"Incorporating Auxiliary Information in Betweenness Measure for Input-Output Networks".
\emph{Physica A: Statistical Mechanics and its Applications}, 607, 128200.
}
