% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra's algorithm}
\usage{
dijkstra(adjmat, src)
}
\arguments{
\item{adjmat}{The adjacency matrix of an directed and weighted network.}

\item{src}{The given source node to find the shortest distance.}
}
\value{
Lists of distance, previous node.
}
\description{
Implement the Dijkstra's algorithm to find the shortest paths
from the source node to all nodes in the given network.
}
\references{
Dijkstra, E. W. (1959).
"A Note on Two Problems in Connexion with Graphs".
\emph{Numerische Mathematik}, 1, 269--271.
}
